requirements.txt
libclang==12.0.0
pytest==6.2.4
tox==3.23.1
Sphinx==4.0.2
sphinx-rtd-theme==0.5.2
coverage==5.5
setuptools==57.0.0
build==0.4.0
twine==3.4.1
sphinx-autodoc-typehints==1.12.0
gitchangelog==3.0.4
setuptools>=42
wheel
